+++ apptpl2/apptpl2_init.h
@@ -1,14 +1,16 @@
 #ifndef INIT_H
 #define INIT_H
 
+#include <stdint.h>
+
 #define DEFAULT_CONFIG_FILE "apptpl2.ini"
 
 
 struct config {
     char version[8];
-    int  http_port;
+    uint16_t  http_port;
     char redis_host[46]; // max 45 bytes, such as 0000:0000:0000:0000:0000:ffff:192.168.100.228
-    int  redis_port;
+    uint16_t  redis_port;
     char redis_passwd[64];
 };
